common_requirements.txt
Django
South
PIL
django-tagging
Fabric==0.9.3
paramiko==1.7.6
pycrypto==2.3
psycopg2==2.3.2

